fix: replace hashtable with native array method
use array of `struct var` in generator to store all visited
identifiers information, such as stack location and their names

hashtable impl was buggy on my part, maybe strdup for ident value,
could have fixed it, but for now don't need it. Because when
we introduce scopes, we may flush `m_vars` generator member, or
flag it for other scope. fitting all the details in a key may,
be tough and prone to more collisions.

TODO: impl generator parse expr with struct var array m_vars