feat: exit with assigned variables value from stack
Assembly will make a copy of the value at register,
that is at an offset of the symbol order declaration.
(yes, its hard for me to describe it as of now, in words)
e.g.
a := 1
b := 2
c := 3
exit(b)

b is at offset of (1 * 8 bytes) from c.
Our stack base pointer will copy the value of register
that holds b's value. so it will pop it and push it to
the top as a copy

Stack trace visualized

a := 1
~ 1 ;; <- base ptr

b := 2
~ 2 ;; <- base ptr
~ 1

c := 3
~ 3 ;; <- base ptr
~ 2
~ 1

exit(b) ;; mov rax, 60
        ;; pop rdi
;; copied by assembler generator instruction,
    ;; in g_gen_stmt() -> g_gen_expr()
    ;;    in scope expr.type === int_lit
~ 2 ;; <- base ptr - copied value - swapped
    ;; push  QWORD  [rsp + 8] ;; offset of (1 * 8 bytes)
~ 3 ;; c is 1 offset from b
~ 2 ;; copied from here
~ 1